requirements.txt
pdf2image
pytesseract
Pillow
pdfplumber
librosa
soundfile
transformers
torch
sentencepiece
nltk
tqdm
rouge-score
